requirements.txt
fastapi
pymongo
celery
python-dotenv
uvicorn
redis
beautifulsoup4
httpx
dishka
